RUMM-633 Update to renamed APIs in Shopist
